+++ BlueprintUI/Sources/Layout/Stack.swift
@@ -379,7 +379,7 @@ extension StackLayout {
         // During layout the constraints are always `.exactly` to fit the provided size
         let vectorConstraint = size.vectorConstraint(axis: axis)
 
-        let frames = _frames(for: items.map(StackLayoutItem.init), in: vectorConstraint)
+        let frames = _frames(for: items.map(StackLayoutItem.init), in: vectorConstraint, scale: UIScreen.main.scale)
 
         return frames.map { frame in
             LayoutAttributes(frame: frame.rect(axis: axis))
@@ -392,7 +392,7 @@ extension StackLayout {
         // During measurement the constraints may be `.atMost` or `.unconstrained` to fit the measurement constraint
         let vectorConstraint = constraint.vectorConstraint(on: axis)
 
-        let frames = _frames(for: items.map(StackLayoutItem.init), in: vectorConstraint)
+        let frames = _frames(for: items.map(StackLayoutItem.init), in: vectorConstraint, scale: UIScreen.main.scale)
 
         let vector = frames.reduce(Vector.zero) { vector, frame -> Vector in
             Vector(
@@ -406,10 +406,11 @@ extension StackLayout {
 
     private func _frames(
         for items: [StackLayoutItem],
-        in vectorConstraint: VectorConstraint
+        in vectorConstraint: VectorConstraint,
+        scale: CGFloat
     ) -> [VectorFrame] {
         // First allocate available space along the layout axis.
-        let axisSegments = _axisSegments(for: items, in: vectorConstraint)
+        let axisSegments = _axisSegments(for: items, in: vectorConstraint, scale: scale)
 
         let axisConstraints = axisSegments.map { $0.magnitude }
 
@@ -428,7 +429,7 @@ extension StackLayout {
     /// sizes along the layout axis, represented as segments.
     ///
     /// The axis segments of a Row look like this diagram.
-    ///
+    /// ```
     /// Row───────────────────────────────────────────┐
     /// │┌───────────┐                                │
     /// ││           │                   ┌───────────┐│
@@ -442,14 +443,15 @@ extension StackLayout {
     /// ││           │                   └───────────┘│
     /// │└───────────┘                                │
     /// └─────────────────────────────────────────────┘
-    ///
+    /// ```
     /// - Parameters:
     ///   - for: The items to measure.
     ///   - in: The constraint for all measurements.
     /// - Returns: The axis measurements as segments.
     private func _axisSegments(
         for items: [StackLayoutItem],
-        in vectorConstraint: VectorConstraint
+        in vectorConstraint: VectorConstraint,
+        scale: CGFloat
     ) -> [Segment] {
 
         let minimumTotalSpacing = CGFloat(items.count - 1) * minimumSpacing
@@ -487,6 +489,7 @@ extension StackLayout {
 
                 measurement.size = measurement.item
                     .measure(in: constraint)
+                    .rounded(.up, by: scale, axis: axis)
             }
 
             // Determine how much space the fixed items took up
@@ -883,7 +886,7 @@ extension StackLayout {
 
         let constraint = proposal.vectorConstraint(on: axis)
 
-        let frames = _frames(for: subelements.map(StackLayoutItem.init), in: constraint)
+        let frames = _frames(for: subelements.map(StackLayoutItem.init), in: constraint, scale: environment.displayScale)
 
         let vector = frames.reduce(Vector.zero) { vector, frame -> Vector in
             Vector(
@@ -903,7 +906,7 @@ extension StackLayout {
     ) {
         let constraint = size.vectorConstraint(axis: axis)
 
-        let frames = _frames(for: subelements.map(StackLayoutItem.init), in: constraint)
+        let frames = _frames(for: subelements.map(StackLayoutItem.init), in: constraint, scale: environment.displayScale)
 
         for i in 0..<subelements.count {
             let vectorFrame = frames[i]
@@ -1251,5 +1254,19 @@ extension ElementBuilder where Child == StackLayout.Child {
 extension StackLayout.Child: ElementBuilderChild {
     public init(_ element: Element) {
         self.init(element: element)
+    }
+}
+
+extension CGSize {
+
+    fileprivate func rounded(_ rule: FloatingPointRoundingRule, by scale: CGFloat, axis: StackLayout.Axis) -> CGSize {
+        switch axis {
+
+        case .horizontal:
+            CGSize(width: width.rounded(rule, by: scale), height: height)
+
+        case .vertical:
+            CGSize(width: width, height: height.rounded(rule, by: scale))
+        }
     }
 }